+++ app/src/main/java/com/meezzi/fingerchoice/explore/ExploreRestaurantFragment.kt
@@ -1,15 +1,15 @@
-package com.meezzi.fingerchoice
+package com.meezzi.fingerchoice.explore
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.meezzi.fingerchoice.databinding.FragmentSavedRestaurantBinding
+import com.meezzi.fingerchoice.databinding.FragmentExploreRestaurantBinding
 
-class SavedRestaurantFragment : Fragment() {
+class ExploreRestaurantFragment : Fragment() {
 
-    private var _binding: FragmentSavedRestaurantBinding? = null
+    private var _binding: FragmentExploreRestaurantBinding? = null
     private val binding get() = _binding!!
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -19,7 +19,7 @@ class SavedRestaurantFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentSavedRestaurantBinding.inflate(inflater, container, false)
+        _binding = FragmentExploreRestaurantBinding.inflate(inflater, container, false)
         return binding.root
     }
