+++ app/src/main/res/layout/fragment_saved_restaurant.xml
@@ -4,6 +4,6 @@
     android:id="@+id/frameLayout4"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SavedRestaurantFragment">
+    tools:context=".save.SavedRestaurantFragment">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
