+++ app/src/main/res/navigation/home_graph.xml
@@ -7,13 +7,13 @@
 
     <fragment
         android:id="@+id/navigation_explore_restaurant"
-        android:name="com.meezzi.fingerchoice.ExploreRestaurantFragment"
+        android:name="com.meezzi.fingerchoice.explore.ExploreRestaurantFragment"
         android:label="ExploreRestaurantFragment"
         tools:layout="@layout/fragment_explore_restaurant" />
 
     <fragment
         android:id="@+id/navigation_map"
-        android:name="com.meezzi.fingerchoice.MapFragment"
+        android:name="com.meezzi.fingerchoice.map.MapFragment"
         android:label="MapFragment"
         tools:layout="@layout/fragment_map" >
         <action
@@ -23,19 +23,19 @@
 
     <fragment
         android:id="@+id/navigation_saved_restaurant"
-        android:name="com.meezzi.fingerchoice.SavedRestaurantFragment"
+        android:name="com.meezzi.fingerchoice.save.SavedRestaurantFragment"
         android:label="SavedRestaurantFragment"
         tools:layout="@layout/fragment_saved_restaurant" />
 
     <fragment
         android:id="@+id/navigation_profile"
-        android:name="com.meezzi.fingerchoice.ProfileFragment"
+        android:name="com.meezzi.fingerchoice.profile.ProfileFragment"
         android:label="ProfileFragment"
         tools:layout="@layout/fragment_profile" />
 
     <fragment
         android:id="@+id/navigation_detailRestaurant"
-        android:name="com.meezzi.fingerchoice.DetailRestaurantFragment"
+        android:name="com.meezzi.fingerchoice.ui.detail.DetailRestaurantFragment"
         android:label="ProfileFragment"
         tools:layout="@layout/fragment_detail_restaurant" />
